New route for login authentication
This route checks first if a user exists based on either an email or username. If the provided string matches with a user's email or username in the database, we move onto authentication.

Authentication is taken care of using the bcrypt library.
Currently we only check if a user could login and respond with an appropriate message.